Added __init__ method to reinitialize file list
This fixes a problem where you do

rep = Report()
(Do stuf with the report)
rep = Report() -- ostensibly to reinitialize the variable

And then find out that the new report contains all the stuff in the
old report. This was due to storing the temp file list as a class
variable which didn't reinitialize. Setting the file list to be an
instance variable which is reinitialized at construction time
fixes this problem.