Switched all plotting to matplotlib "headless" mode
The problem was that the many successive calls to the display server
occurring when generating reports for many models on orchestra was
leading to crashes. Now figure objects are created and managed
explicitly, avoiding calls to the interactive API.

This required changes to:
- plot_data and fit_plotting_function in nbd_mcmc_pysb
- plotting routines in experiment.py and knowledge.py

I also made corresponding changes in bayessb.