Begin rewrite of grid_analysis using pandas
Started with a reimplementation of the plot_timecourse function. Still
can't figure out why I can't take a cross section of the data using the
liposome axis, so right now it just works with a single lipo
concentration.

Added a test to make sure the plot_timecourses works, but nose isn't
working with the current version of numpy (1.7) which I just upgraded
to. Arggh.