Moved generic elements of NBD_MCMC to a new class
The new class, tbidbaxlipo.mcmc.MCMC, contains the enhancements to the
underlying bayessb.MCMC that are useful for working with builder-based
models with complex observables, but aren't specific to NBD fluorescence
modeling.

The idea was to extract these features so as to build a new class for
fitting dye release data.