Generic implementation of priors for NBD fitting
Involved the creation of classes for prior distributions (Normal,
Uniform), and revision of models.core to have new implementations of
the functions prior and random_initial_values to use the new list of
priors.

Also modified the model creation in nbd_multiconf to include uniform
prior probability distributions on the scaling and rate parameters.