Tracked down an error in histogram normalization
in which I was using the trapezoidal rule, which was leading to
distortions of the normalized frequencies of distributions with small
averages. Instead I now normalize by simply dividing by the total
frequency, which is significantly better.

Also, I moved the job simulation code of the package code for the
translocation module.