Updated HDF5 saving script and timecourse plotting
Changed it so that the final dimension in the HDF5 file is time,
so that the : can be omitted (since most plots, etc.) are over
all time.

Updated the code in plot_timecourse.py (formerly __init__.py) to use the saved
HDF5 file.

Note that this is the commit that was used to generate the 300 * 4 = 1200
simulations in the data directories.