Code, simulation data, and plot func for tbid titration
tbid titration for a simple model consisting of tBid/Bax translocation, Bax
activation, and pores from Bax monomers. 30 x 5 = 150 simulations for each of
21 concentrations (0-20 nM tBid). Interestingly, the plots show that the ODE
model is fundamentally different from the stochastic model, in that the
stochastic model shows the steady-state Fmax going up but the kinetics fairly
consistent, whereas the deterministic model shows the Fmax always at 1.0, and
the kinetics increasing with increasing tBid.