Made likelihood, not get_likelihood_function, default
Previously I had written closures around the likelihood function using a
method get_likelihood_function (as in the current implementation of
nbd_mcmc). This was the default behavior, and as such, the setstate
method in mcmc.MCMC looked there for the likelhood function. However,
with the advent of pysb expressions, there is less need for closures
around the likelihood function. As such, I'm keeping this approach for
nbd_mcmc but am making self.likelihood (a non closure) the default
behavior. This requires changing the setstate method in mcmc.MCMC and
adding an overriden setstate method to nbd_mcmc.