Edited the data for row B for Triton case
Since I forgot to add Triton when measuring the baseline (no DPX)
condition, I decided to use the Triton wells in row F as my no dilution
values. The easiest way to do this was to simply copy the values over
from row F to row B. I put these modified version in a separate file,
with the suffix _fixed.