Begin purging compartment syntax
I finally concluded that the use of compartments, while convenient for
stochastic multicompartmental simulation in BNG, was not only not useful
for simulations with Kappa, but added to significant confusion in the
formulation of the models. For example, the localization of Bid or Bax
could be specified using the "loc" keyword and/or by the compartment
status.

In addition, I finally deemed it time to purge the overloaded site "loc"
for Bid and Bax in favor of separate sites "cpt", denoting the
compartment (solution or vesicles), "conf" for conformation (eg. aqueous
conformation, or membrane conformation 1 or 2, etc.) and "pore" for
whether the Bax is in a pore or not.

My idea is to ultimately eliminate the n_cpt approach to
compartmentalization, which uses compartmental BNGL, in favor of a
site-based approach to compartments similar to what is implemented in
site_cpt; except that I hope to handle the site-based
compartmentalization in a fashion that doesn't require all of the
interaction macros to be rewritten. I believe that this should be doable
by creating a one_cpt (single compartment) model and then from that,
generating a new, multi-compartmental model by post-processing,
generating additional rules for the different compartments as necessary
and adjusting rates.

To aid in this process, I introduced some tests that will make sure that
each of the build_model macros in core.py can successfully build the
respective model. This should insure that I haven't missed any glaring
syntactical or rule-level errors.

Wish me luck.