Progress on conversion of one to multi cpt models
Basic functionality in place in multi_cpt.make_multi_compartment. For
safety, need tests to make sure that the references to the monomers and
parameters within the complex patterns of observables, initial
conditions, rules, etc. have the proper integrity. Am in the middle of
writing a test that will guarantee this.

Major things remaining to be done:

* First, the number of vesicles, and the scaling factor need to be
  specified in the builder.

* Second, the make_multi_compartment function needs to use the
  compartment number to generate the appropriate number of monomers,
  rules, etc.

* Third, the parameters will need to be scaled accordingly, and

* Fourth, the rules for translocation to and from vesicles will need to
  be changed so that they do not use a pseudo-first order reaction with
  the Vesicles() monomer.