Code and doc script to output .csv files of fit values
I created a new class specific to bid_bim_nbd_release, FitResult, that
stores the results of fit values or initial slope calculations. In this
format it's easy to return the results from the plot scripts and then
convert them to .csv.

Ultimately, it would be better to have the fitting not occur twice, once
for the fitting and once for the plotting. However, that would require
the iPython directive to be able to output plots more robustly.