Moved multi_cpt to site_cpt
And moved the previous site_cpt to old_site_cpt.

Since multi_cpt (the new, universal compartmentalization implementation)
is site-based, it makes sense to call it site_cpt, at least for now.
The only thing is I want to make it so that the simulation class
simulates it using BNG instead of Kappa, because that seems to be
faster.