Added option to prevent pore Bax from retrotranslocating
The problem was how to implement the Schwarz model (which lacks a
separate Bax activation step) with irreversible pore formation. One
approach would have been to have the Bax at the membrane simply convert
into a pore. However, this would have involved the creation of a
separate pore function and the connection with reversibility of membrane
binding would have been lost. Instead, I modified the translocation
function to allow the case in which translocation away from the membrane
is conditional on the Bax not being in a pore, which should have the
same effect.