Iterative mean and sd calculations for HDF5 datasets
The numpy implementation of mean involves loading the entire HDF5
dataset into memory, which can be on the order of gigabytes. The SD
calculation actually loads two copies of the array (because it
internally performs the mean calculation). This can lead to memory and
performance issues, particularly when running on a laptop. With Jeremy
Muhlich I wrote versions of the mean and SD calculations that iterate
over slices of the arrays and fill in the mean/sd values along the way,
which reduces memory requirements to negligible levels. However, for now
I've refrained from making these the default implementation, because on
Orchestra the machines have up to 96 gb of RAM, which makes the original
implementation more efficient.