Change to ves_agg that breaks rate=0, fixes inf rate
That is to say, changing the translocation rate in the way shown causes
the model, at the highest agg rates, to match the behavior of the
ves_schwarz_agg_inf model. However, in doing so it also fails to match
the behavior of the multi-compartment simulation when the agg_rate is
zero.