Basic competitive binding, MCMC
Actually can get fits from this. However, I believe this method relies
on the approximation that Lf ~= L0 (i.e., there's a very small amount of
the receptor so that the ligand is not appreciable depleted). Would be
better to have a version in which this assumption is not made.