Script to generate model ensemble fit files
The script takes a yaml file where the implementations of each feature are
defined as a list, then takes the cross product over all features and generates
a list of .yaml files, each of which contains the instructions for a single
fit. The individual fitting runs then get their fit information from these
.yaml files.

Currently not complete because parameters differ from model to model: need
ability to specify that all parameters should be global, so that individual
parameter names do not need to be specified.