Build .mcmc from dynamic list of .yaml files
Work in progress. With this commit comes an update to generate_model_... that,
in addition to generating individual .yaml files for every model fit, also
creates a dependency list in a text file that can then be imported in the
Makefile. Right now, with dummy targets, this appears to work.

The next step is to prevent submitting new MCMC fit jobs when the actual hash
of the .yaml fit file hasn't changed.