Added ability to fit time offset
If specified in the fit YAML file, the time offset will be added to the model
as an additional parameter, which will then be used in the likelihood function
to add a single addition timepoint prior to 0. This additional timepoint is
then ignored when comparing the simulation output to data.