Updated dataset from Justin
From Justin's email, 2/13/2015:

"Attached is the RIGHT data for 126C and 54C. I looked into it and I
don't know what I was thinking on December 3rd, but for the 126C, I sent
you the moving average of 5 points instead of the absolute calculations
for FRET at each time point. Apologies!
Attached is the proper data."