Separate NBD/Release endpt plots, make WT norm optional
As it turns out not all of the datasets appear to have a WT release
control, which rules out normalization to WT. In the meantime, I've
made normalization to WT optional. To make this less confusing, I've
separated out the NBD F/F0 plot func from the release endpt plot func.