New impl of initial rate calculations
Moved the approach in which initial rates were calculated by fitting to
a line to plot_initial_rate_fits in nbd_bax_analysis.py; note that some
plots of ratios were not transferred into the new version--if you want
to see them, look at the deleted code from
bid_bim_nbd_release/bid_bim_nbd_release.py. I also created a new
function that calculates initial rates based on values at early
timepoints.

From an email to Justin about this approach:

I'm working on a new plot I think you'll like. I've been trying to
figure out how to show the NBD vs. Tb initial rates; instead of plotting
the ratios in a bar plot, as I'd tried before, this time tried making a
scatter plot instead.

I decided that getting the initial rate by fitting the first few points
to a line was not actually the best way, since it resulted in fits to
release curves that had a negative intercept, which doesn't make sense;
and even though there was fairly little noise in the release data, the
noise in the initial rate was artifactually increased in the fit by the
curvature in some replicates vs. others (the lag phase). Similarly for
NBD: NBD was even more sensitive to nonlinearities than the release
curve because for many residues it starts to plateau really fast, so the
linear fit doesn't work so well.

So instead of fitting with a line, I just took the release and NBD
values at an arbitrary time point (the 4th timepoint, ~106 seconds). The
variability in the data is then simply established by taking the mean/SD
over the three replicates. I then plotted the NBD/Tb values on a scatter
plot. The results of this are in the "scatter_no_norm" plot attached.
(Please excuse figure cosmetics for now). Note that the vertical dotted
line gives the release value at 100 seconds for wildtype as a reference.

The thing about the "no_norm" plot is that the NBD rates aren't
normalized to their maxes, as we discussed; the raw F/F0 value at 100
seconds is not that meaningful if one is one its way to a max F/F0 of
1.5 and another is on its way to F/F0 of 5.

So for the second plot, "scatter_norm", I normalized the F/F0 at 100
seconds by the maximum attained by that mutant (regardless of whether it
occurs at an intermediate time or the endpoint). Note that this kind of
normalization is what your brain does implicitly when comparing the NBD
curves for different mutants on different plot scales.

When you look at the scatter plot, now most of residues fall roughly on
a line, which makes sense--for many residues, the addition of the dye at
a particular spot probably destabilizes the protein and results in both
faster dye release and faster relative NBD signals.

What's striking here are the outliers: 40, 54, 62, 68, and 79. These
residues all have much faster changes in NBD relative to their dye
release (as we've already observed qualitatively). They are also mostly
slower than wild type, except for 54, which is a bit faster. I grouped
54-79 as "BH3" region residues and color-coded them red so it would be
clearer that these are the residues that define the outliers. For
comparison, I colored the residues in helix 9 green.

The linear fit (blue line) is made to the residues *after removing* 40,
54, 62, 68, and 79 as outliers.

I've also got a plot where I fit the data (after removing outliers) with
a line forcing an intercept at 0, but I'm not sure that's the right
approach, since in theory it's possible to have an increase in NBD but
not a proportional increase in dye release--the lag phase in fact
suggests that this will be the case.