New data for Bid/Bim FRET dataset, now with WT
From email 2-27-15 from Justin:

"Here it is, I did not input any data for the DAC or NBD channels since
there is no FRET happening with DAC-Bid or Dac-Bim and no NBD labeled
Bax here.

It is appended at the end after 188C. Just FYI, these reps are across
the NBD Bax data set. Rep 1 was done along with 3C, rep 2 was done along
with 126C and rep 3 was done along with 184C."

And then my response, 3-6-15:

"FYI I edited the spreadsheet to put WT first so that I could keep a
consistent order with the other dataset. The dataset I'm using is
attached, and I've also put it in the dropbox."