Temp fix of copy error in the Bid/Bim FRET dataset
My email to Justin, 3-6-15:

"Uh-oh! One more wrinkle. As it turns out several of the WT reps have
the words "Time" and "Tb:DPA" as their first points. This copy error was
in the original you sent me and I didn't see it when I moved the
columns.

I could just delete those entries and move the timecourses up one row,
but then those timecourses are short one datapoint and not all of them
are the same length anymore. Do you think you can go back to the
originals and fix?"

---

Note that my fix for now was to delete the text, move the timecourses up
and then duplicate the final points for the affected timecourses. This
way I don't have to modify the code for when the fixed data comes in.