Final tweaks to bid_bim_nbd_release before deleting
This includes addition of filtering and removal of moving average step
which is already reflected in the updated version in nbd_bax_analysis.
This version does, however, include the calculations of peak NBD and
release slopes which were taken out of the final version because
ultimately it didn't seem that relevant.