Removed last five points from dataset 2
Email exchange with Justin:

John -----

Hi Justin,

Another thing: with the second dataset (Bid/Bim FRET + NBD + Tb) there
are some strange outliers at the final points for some of the
timecourses. For example, go to

http://sorger.med.harvard.edu/data/bachman/kale/bid_bim_fret_nbd_release/raw_data.html

And scroll down to 138C. Both Rep 2, for both Release/NBD and cBid/Bim
have big swings for the last few points. Not sure what the reason for
this is, but it tends to make some of the endpoint plots all screwy. For
now, I think I'll just chop off the last 5 points or so of all the
curves.

There are other major outliers in these curves that are not at the
endpoints that we'll probably want to deal with, but we'll deal with
that when we get to it.

Justin -----

Those final points are likely due to when I add detergent (chaps)  to
get full release of Tb:dpa. The signal jumps for every wavelength after
adding chaps.  For all the data sets,  I grouped them in excel and just
copy/pasted formulas for the sheets.  So the last time points are
different for every trace depending on the day I did them.  Cutting off
the last 5 data points is probably the best idea,  this makes every
trace end at the same time point and eliminates the jump.