Save state during burn-in
Saves the current position and the state of the global random number
generator (which is what is used by PTSampler--it does not use the
internal sampler.random_state property, for some reason. Actually this
makes things easier for us). The position filename to use is given in
the call to run_pt (and the Makefile has been modified accordingly). If
the file does not exist, it is created during the first save. If it does
exist, the position is used for the start position.

Note that if the number of walkers, temperatures, or temperature
spacings ever changes, these position files must be replaced.