pt_sample won't run longer than the given burn steps
Previously, it ran convergence_interval numbers of steps, even if this
number (e.g. 50) was greater than the specified number of burn steps.
This was usually OK because the number of burn steps is obviously much
larger than the number of burn-in steps. However, for quick tests to see
if everything is running, it's useful to specify just a few steps, even
1. This makes it so it will only run one step if desired.

Note however that right now, even if burn_steps is 1, it still runs two
burn-in steps: one with num_iterations=1, and one with num_iterations=0.
Need to fix this so it runs the correct number of steps.