one_cpt and lipo_sites sims are approximately equal
when the number of sites is large relative to the amount of Bax.

Note that this requires the forward translocation rate constant to be
scaled by the number of sites per liposome in order for it to be
nominally equivalent to the rate constant used in the one_cpt case.